requirements.txt
Babel==2.8.0
chardet==3.0.4
decorator==4.4.2
docutils==0.16
ebaysdk==2.2.0
feedparser==5.2.1
gevent==20.5.1; sys_platform == 'win32'
html2text==2020.1.16
Jinja2==2.11.2
libsass==0.20.0
lxml ; sys_platform == 'win32'
Mako==1.1.2
MarkupSafe==1.1.1
mock==4.0.2
num2words==0.5.10
ofxparse==0.20
passlib==1.7.2
Pillow==7.1.2; python_version < '3.7' or sys_platform != 'win32'
polib==1.1.0
psutil==5.7.0
psycopg2==2.8.5; sys_platform == 'win32' or python_version >= '3.8'
pydot==1.4.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
pyusb==1.0.2
qrcode==6.1
reportlab==3.5.42
requests==2.23.0
zeep==3.4.0
vatnumber==1.2
vobject==0.9.6.1
Werkzeug==0.16.0
XlsxWriter==1.2.8
xlwt==1.3.0
xlrd==1.2.0
pypiwin32 ; sys_platform == 'win32'
simplejson==3.17.0
Flask~=1.1.2
googlemaps~=4.4.2
geopy~=2.0
humanize==3.1.0
phonenumbers
numpy
sklearn
matplotlib
firebase_admin
pyfcm
matrix_client